Create new stacking context for link highlights
Alec mentioned wanting to avoid entering z-index hell to fix this.
Creating a new stacking context is a bit better, but still sucks because
we have to do this on every parent element.

There should be a way for us to create a link without highlighting it
using a psuedo element position -1 in the z direction. I'll need to play
with it.

Fixes #12