Use background and overflow for link highlights
This is an alternative method to the blue highlight we apply to links.
Instead of using a pseudo element position at z=-1 on all links, we set
a fixed height and allow text to overflow out of the container.

This isn't a pixel perfect solution. The current downside of this is
that links are 1px further down than they should be. However, it does
fix several issues:

1. We no longer need to create a new stacking context for links not on
   the root context.
2. CORE is now properly highlighted under Taylor's bio. We previously
   had a small blue block with no text around it.
3. Padding is more consistent. (See my profile page with the links to
   HuskyTrails and TIP for comparison.)