Upgrade to webpack 4 to fix Uglify build errors
The following error began appearing on CI in the last week.

    ERROR in scripts/main.bundle.js from UglifyJs
    Unexpected token: punc (() [scripts/main.bundle.js:1538,7]

Colin traced this error and found that it was due to lack of ES2015
support by the built-in UglifyJS minifier. There's two ways this can be
fixed.

  1. By installing the ES2015 version of UglifyJs and using that instead
     of the version built into webpack 3.
  2. Upgrade to webpack 4, which has a newer version of UglifyJS
     already.

Approach #2 was taken since it's a good idea to stay up to date with
dependencies anyway. A few other changes were made while updating to
webpack 4.

  - extract-text-webpack-plugin was swapped out for
    mini-css-extract-plugin since the former was not upgraded for
    webpack 4 yet. See: https://goo.gl/bXDK5S
  - vue-loader was upgraded in package.json for webpack 4 compatibility.
  - html-loader was also upgraded for webpack4 compatibility.

SW-10 #close
SW-10 #time 1h 15m