Separate link_ref into href and @click props
There are many reasons why href and @click work better as separate props
compared to a single one that's type inspected on the callee end.

  1. Some links may both direct a user to another page while
     simultaneously having some kind of JavaScript side-effect.
  2. The ommision of href tells us that we probably have a button rather
     than an anchor component.
  3. @click is an existing pattern that's commonly used in Vue.

The changes here also enable:

  1. The web browser to know where a link will go before clicking on it.
     This brings back the standard location indicator (typically on the
     bottom left).
  2. Us to inherit all of the goodies of vue-router's router-link, since
     it's now used under the hood. (Ex: active styling)

The feature to specify integers as arguments to $router.go() was
removed. There weren't any parts of the site that used it and for any
future links that do, a simple @click="$router.go(-1)" will suffice.

Fixes #70