Restore page transition out animation
This got removed accidentally during the transition to Nuxt.js. Nuxt.js
has a default page transition with the name "page" and mode "out-in". We
just have to rename our page animation CSS class to take advantage of
it.

I'm guessing the reason the current method (composing <nuxt /> with
<transition>) didn't work is because there's only ever one <nuxt />
component.

https://nuxtjs.org/api/configuration-transition/

Fixes #72